+++ feature/src/adm_tools.h
@@ -17,6 +17,7 @@
  */
 
 #define M_PI 3.14159265358979323846264338327
+#include <math.h>
 #include "common/macros.h"
 
 #pragma once