+++ feature/src/psnr_tools.c
@@ -17,6 +17,7 @@
  */
 
 #include <stdio.h>
+#include <string.h>
 #include "psnr_tools.h"
 
 /*