+++ feature/src/vmaf_main.c
@@ -26,6 +26,7 @@
 int adm(int (*read_frame)(float *ref_data, float *main_data, float *temp_data, int stride, void *user_data), void *user_data, int w, int h, const char *fmt);
 int ansnr(int (*read_frame)(float *ref_data, float *main_data, float *temp_data, int stride, void *user_data), void *user_data, int w, int h, const char *fmt);
 int vif(int (*read_frame)(float *ref_data, float *main_data, float *temp_data, int stride, void *user_data), void *user_data, int w, int h, const char *fmt);
+int vifdiff(int (*read_frame)(float *ref_data, float *main_data, float *temp_data, int stride, void *user_data), void *user_data, int w, int h, const char *fmt);
 int motion(int (*read_noref_frame)(float *main_data, float *temp_data, int stride, void *user_data), void *user_data, int w, int h, const char *fmt);
 int all(int (*read_frame)(float *ref_data, float *main_data, float *temp_data, int stride, void *user_data), void *user_data, int w, int h, const char *fmt);
 
@@ -39,6 +40,7 @@ static void usage(void)
          "\tansnr\n"
          "\tmotion\n"
          "\tvif\n"
+         "\tvifdiff\n"
          "\tall\n"
          "fmts:\n"
          "\tyuv420p\n"
@@ -127,6 +129,8 @@ fail_or_end_noref:
             ret = vif(read_frame, s, w, h, fmt);
         else if (!strcmp(app, "all"))
             ret = all(read_frame, s, w, h, fmt);
+        else if (!strcmp(app, "vifdiff"))
+            ret = vifdiff(read_frame, s, w, h, fmt);
         else
             ret = 2;