+++ python/src/vmaf/core/feature_assembler.py
@@ -4,6 +4,7 @@ __license__ = "Apache, Version 2.0"
 from vmaf.core.feature_extractor import FeatureExtractor
 from vmaf.core.result import BasicResult
 from vmaf.core.executor import run_executors_in_parallel
+from vmaf.core.matlab_feature_extractor import StrredFeatureExtractor, SpEEDMatlabFeatureExtractor
 
 class FeatureAssembler(object):
     """