+++ python/src/vmaf/core/feature_extractor.py
@@ -269,6 +269,84 @@ class VmafFeatureExtractor(FeatureExtractor):
 
         return result
 
+class VifFrameDifferenceFeatureExtractor(FeatureExtractor):
+
+    TYPE = "VifDiff_feature"
+
+    VERSION = '0.1'
+
+    ATOM_FEATURES = ['vifdiff',
+                     'vifdiff_num', 'vifdiff_den',
+                     'vifdiff_num_scale0', 'vifdiff_den_scale0',
+                     'vifdiff_num_scale1', 'vifdiff_den_scale1',
+                     'vifdiff_num_scale2', 'vifdiff_den_scale2',
+                     'vifdiff_num_scale3', 'vifdiff_den_scale3',
+                     ]
+
+    DERIVED_ATOM_FEATURES = ['vifdiff_scale0', 'vifdiff_scale1', 'vifdiff_scale2', 'vifdiff_scale3',
+                             ]
+
+    ADM2_CONSTANT = 0
+    ADM_SCALE_CONSTANT = 0
+
+    def _generate_result(self, asset):
+        # routine to call the command-line executable and generate feature
+        # scores in the log file.
+
+        quality_width, quality_height = asset.quality_width_height
+        log_file_path = self._get_log_file_path(asset)
+
+        yuv_type=self._get_workfile_yuv_type(asset)
+        ref_path=asset.ref_workfile_path
+        dis_path=asset.dis_workfile_path
+        w=quality_width
+        h=quality_height
+        logger = self.logger
+
+        ExternalProgramCaller.call_vifdiff_feature(yuv_type, ref_path, dis_path, w, h, log_file_path, logger)
+
+    @classmethod
+    def _post_process_result(cls, result):
+        # override Executor._post_process_result
+
+        result = super(VifFrameDifferenceFeatureExtractor, cls)._post_process_result(result)
+
+        # vifdiff_scalei = vifdiff_num_scalei / vifdiff_den_scalei, i = 0, 1, 2, 3
+        vifdiff_num_scale0_scores_key = cls.get_scores_key('vifdiff_num_scale0')
+        vifdiff_den_scale0_scores_key = cls.get_scores_key('vifdiff_den_scale0')
+        vifdiff_num_scale1_scores_key = cls.get_scores_key('vifdiff_num_scale1')
+        vifdiff_den_scale1_scores_key = cls.get_scores_key('vifdiff_den_scale1')
+        vifdiff_num_scale2_scores_key = cls.get_scores_key('vifdiff_num_scale2')
+        vifdiff_den_scale2_scores_key = cls.get_scores_key('vifdiff_den_scale2')
+        vifdiff_num_scale3_scores_key = cls.get_scores_key('vifdiff_num_scale3')
+        vifdiff_den_scale3_scores_key = cls.get_scores_key('vifdiff_den_scale3')
+        vifdiff_scale0_scores_key = cls.get_scores_key('vifdiff_scale0')
+        vifdiff_scale1_scores_key = cls.get_scores_key('vifdiff_scale1')
+        vifdiff_scale2_scores_key = cls.get_scores_key('vifdiff_scale2')
+        vifdiff_scale3_scores_key = cls.get_scores_key('vifdiff_scale3')
+        result.result_dict[vifdiff_scale0_scores_key] = list(
+            (np.array(result.result_dict[vifdiff_num_scale0_scores_key])
+             / np.array(result.result_dict[vifdiff_den_scale0_scores_key]))
+        )
+        result.result_dict[vifdiff_scale1_scores_key] = list(
+            (np.array(result.result_dict[vifdiff_num_scale1_scores_key])
+             / np.array(result.result_dict[vifdiff_den_scale1_scores_key]))
+        )
+        result.result_dict[vifdiff_scale2_scores_key] = list(
+            (np.array(result.result_dict[vifdiff_num_scale2_scores_key])
+             / np.array(result.result_dict[vifdiff_den_scale2_scores_key]))
+        )
+        result.result_dict[vifdiff_scale3_scores_key] = list(
+            (np.array(result.result_dict[vifdiff_num_scale3_scores_key])
+             / np.array(result.result_dict[vifdiff_den_scale3_scores_key]))
+        )
+
+        # validate
+        for feature in cls.DERIVED_ATOM_FEATURES:
+            assert cls.get_scores_key(feature) in result.result_dict
+
+        return result
+
 class PsnrFeatureExtractor(FeatureExtractor):
 
     TYPE = "PSNR_feature"
@@ -453,97 +531,3 @@ class MsSsimFeatureExtractor(FeatureExtractor):
         ExternalProgramCaller.call_ms_ssim(yuv_type, ref_path, dis_path, w, h, log_file_path, logger)
 
 
-class MatlabFeatureExtractor(FeatureExtractor):
-
-    @classmethod
-    def _assert_class(cls):
-        # override Executor._assert_class
-        super(MatlabFeatureExtractor, cls)._assert_class()
-        VmafExternalConfig.get_and_assert_matlab()
-
-class StrredFeatureExtractor(MatlabFeatureExtractor):
-
-    TYPE = 'STRRED_feature'
-
-    # VERSION = '1.0'
-    # VERSION = '1.1' # fix matlab code where width and height are mistakenly swapped
-    VERSION = '1.2' # fix minor frame and prev frame swap issue
-
-    ATOM_FEATURES = ['srred', 'trred', ]
-
-    DERIVED_ATOM_FEATURES = ['strred', ]
-
-    MATLAB_WORKSPACE = VmafConfig.root_path('matlab', 'strred')
-
-    @classmethod
-    def _assert_an_asset(cls, asset):
-        super(StrredFeatureExtractor, cls)._assert_an_asset(asset)
-        assert asset.ref_yuv_type == 'yuv420p' and asset.dis_yuv_type == 'yuv420p', \
-            'STRRED feature extractor only supports yuv420p for now.'
-
-    def _generate_result(self, asset):
-        # routine to call the command-line executable and generate quality
-        # scores in the log file.
-
-        ref_workfile_path = asset.ref_workfile_path
-        dis_workfile_path = asset.dis_workfile_path
-        log_file_path = self._get_log_file_path(asset)
-
-        current_dir = os.getcwd() + '/'
-
-        ref_workfile_path = make_absolute_path(ref_workfile_path, current_dir)
-        dis_workfile_path = make_absolute_path(dis_workfile_path, current_dir)
-        log_file_path = make_absolute_path(log_file_path, current_dir)
-
-        quality_width, quality_height = asset.quality_width_height
-
-        strred_cmd = '''{matlab} -nodisplay -nosplash -nodesktop -r "run_strred('{ref}', '{dis}', {h}, {w}); exit;" >> {log_file_path}'''.format(
-            matlab=VmafExternalConfig.get_and_assert_matlab(),
-            ref=ref_workfile_path,
-            dis=dis_workfile_path,
-            w=quality_width,
-            h=quality_height,
-            log_file_path=log_file_path,
-        )
-        if self.logger:
-            self.logger.info(strred_cmd)
-
-        os.chdir(self.MATLAB_WORKSPACE)
-        run_process(strred_cmd, shell=True)
-        os.chdir(current_dir)
-
-    @classmethod
-    def _post_process_result(cls, result):
-        # override Executor._post_process_result
-
-        def _strred(srred_trred):
-            srred, trred = srred_trred
-            try:
-                return srred * trred
-            except TypeError: # possible either srred or trred is None
-                return None
-
-        result = super(StrredFeatureExtractor, cls)._post_process_result(result)
-
-        # calculate refvar and disvar from ref1st, ref2nd, dis1st, dis2nd
-        srred_scores_key = cls.get_scores_key('srred')
-        trred_scores_key = cls.get_scores_key('trred')
-        strred_scores_key = cls.get_scores_key('strred')
-
-        srred_scores = result.result_dict[srred_scores_key]
-        trred_scores = result.result_dict[trred_scores_key]
-
-        # compute strred scores
-        # === Way One: consistent with VMAF framework, which is to multiply S and T scores per frame, then average
-        # strred_scores = map(_strred, zip(srred_scores, trred_scores))
-        # === Way Two: authentic way of calculating STRRED score: average first, then multiply ===
-        assert len(srred_scores) == len(trred_scores)
-        strred_scores = ListStats.nonemean(srred_scores) * ListStats.nonemean(trred_scores) * np.ones(len(srred_scores))
-
-        result.result_dict[strred_scores_key] = strred_scores
-
-        # validate
-        for feature in cls.DERIVED_ATOM_FEATURES:
-            assert cls.get_scores_key(feature) in result.result_dict
-
-        return result