+++ python/src/vmaf/core/result.py
@@ -14,7 +14,7 @@ __license__ = "Apache, Version 2.0"
 class BasicResult(object):
     """
     Has some basic functions, but don't need asset or executor_id. To be used by
-    FeatureAssemler, which is not an Executor.
+    FeatureAssembler, which is not an Executor.
     """
     def __init__(self, asset, result_dict):
         self.asset = asset
@@ -49,17 +49,48 @@ class BasicResult(object):
         a corresponding list of scores that uses key '*_scores'. For example,
         if the key is 'VMAF_score', there must exist a corresponding key
         'VMAF_scores'.
+        The list of scores (when present) should also follow the convention
+        that IF multiple models are present (e.g. in bootstrapping),
+        the first dimension will be the model and the
+        second dimension the video frames. This means that a 2D result array
+        should be created ONLY when multiple models are present.
         """
         if re.search(r"_score$", key):
             scores_key = self.get_scores_key_from_score_key(key)
             if scores_key in self.result_dict:
                 scores = self.result_dict[scores_key]
-                return self.score_aggregate_method(scores)
+                # if scores are in a list, wrap in an array to be consistent
+                if type(scores) is list:
+                    scores = np.asarray(scores)
+                # dimension assertion: scores should be either 1-D (one prediction per frame) or 2-D (single/multiple predictions per frame)
+                assert scores.ndim <= 2, 'Per frame score aggregation is not well-defined; scores cannot be saved in a N-D array with N > 2.'
+
+                # check if there are more than one models (first dimension)
+                if scores.ndim == 2:
+                    # check that there are > 1 models present
+                    assert scores.shape[0] > 1, '# models is <=1, but a 2D result array (models x frames) was used.'
+                    # check that there are >= 1 frames predicted
+                    assert scores.shape[1] >= 1, '# predicted frames is < 1.'
+                    # apply score aggregation on each individual model
+                    # a scores "piece" corresponds to a single model's predictions over all frames
+                    return [self.score_aggregate_method(scores_piece) for scores_piece in scores]
+                else:
+                    # just one prediction per frame
+                    return self.score_aggregate_method(scores)
         raise KeyError(error)
 
     def get_ordered_list_scores_key(self):
         # e.g. ['VMAF_scores', 'VMAF_vif_scores']
-        list_scores_key = filter(lambda key: re.search(r"_scores$", key),
+        # filter out scores that are > 1D (e.g. when having multiple models in bootstrapping)
+        list_scores_key = filter(lambda key: re.search(r"_scores$", key) and np.asarray(self.result_dict[key]).ndim == 1,
+                                 self.result_dict.keys())
+        list_scores_key = sorted(list_scores_key)
+        return list_scores_key
+
+    def get_ordered_list_multimodel_scores_key(self):
+        # e.g. ['BOOTSTRAP_VMAF_all_models_scores']
+        # select only scores that are > 1D (e.g. when having multiple models in bootstrapping)
+        list_scores_key = filter(lambda key: re.search(r"_scores$", key) and np.asarray(self.result_dict[key]).ndim > 1,
                                  self.result_dict.keys())
         list_scores_key = sorted(list_scores_key)
         return list_scores_key
@@ -67,6 +98,11 @@ class BasicResult(object):
     def get_ordered_list_score_key(self):
         # e.g. ['VMAF_score', 'VMAF_vif_score']
         list_scores_key = self.get_ordered_list_scores_key()
+        return map(lambda scores_key: scores_key[:-1], list_scores_key)
+
+    def get_ordered_list_multimodel_score_key(self):
+        # e.g. ['BOOTSTRAP_VMAF_all_models_score']
+        list_scores_key = self.get_ordered_list_multimodel_scores_key()
         return map(lambda scores_key: scores_key[:-1], list_scores_key)
 
     def _get_scores_str(self, unit_name='Frame'):
@@ -201,6 +237,20 @@ class Result(BasicResult):
         list_score_key = self.get_ordered_list_score_key()
         list_scores = map(lambda key: self.result_dict[key], list_scores_key)
         list_aggregate_score = map(lambda key: self[key], list_score_key)
+
+        list_multimodel_scores_key = self.get_ordered_list_multimodel_scores_key()
+        list_multimodel_score_key = self.get_ordered_list_multimodel_score_key()
+        # here we need to transpose, since printing is per frame and not per model
+        # we also need to turn the 2D array to a list of lists, for unpacking to work as expected
+        list_multimodel_scores = map(lambda key: self.result_dict[key].T.tolist(), list_multimodel_scores_key)
+        list_aggregate_multimodel_score = map(lambda key: self[key], list_multimodel_score_key)
+
+        # append multimodel scores and keys (if any)
+        list_scores_key += list_multimodel_scores_key
+        list_score_key += list_multimodel_score_key
+        list_scores += list_multimodel_scores
+        list_aggregate_score += list_aggregate_multimodel_score
+
         list_scores_reordered = zip(*list_scores)
 
         def prettify(elem):
@@ -349,6 +399,20 @@ class Result(BasicResult):
         list_score_key = self.get_ordered_list_score_key()
         list_scores = map(lambda key: self.result_dict[key], list_scores_key)
         list_aggregate_score = map(lambda key: self[key], list_score_key)
+
+        list_multimodel_scores_key = self.get_ordered_list_multimodel_scores_key()
+        list_multimodel_score_key = self.get_ordered_list_multimodel_score_key()
+        # here we need to transpose, since printing is per frame and not per model
+        # we also need to turn the 2D array to a list of lists, for unpacking to work as expected
+        list_multimodel_scores = map(lambda key: self.result_dict[key].T.tolist(), list_multimodel_scores_key)
+        list_aggregate_multimodel_score = map(lambda key: self[key], list_multimodel_score_key)
+
+        # append multimodel scores and keys (if any)
+        list_scores_key += list_multimodel_scores_key
+        list_score_key += list_multimodel_score_key
+        list_scores += list_multimodel_scores
+        list_aggregate_score += list_aggregate_multimodel_score
+
         list_scores_reordered = zip(*list_scores)
 
         top = OrderedDict()