+++ wrapper/src/combo.c
@@ -34,6 +34,7 @@
 #include "adm_options.h"
 #include "combo.h"
 #include "debug.h"
+#include "psnr_tools.h"
 
 #ifdef MULTI_THREADING
 #include "common/blur_array.h"
@@ -852,7 +853,10 @@ int combo(int (*read_frame)(float *ref_data, float *main_data, float *temp_data,
 
     // start threads
     int t;
-    pthread_t thread[combo_thread_data.thread_count];
+	int numThread = combo_thread_data.thread_count;
+	pthread_t* thread = (pthread_t*)calloc(numThread, sizeof(pthread_t));
+	memset(thread, 0, numThread * sizeof(pthread_t));
+
     for (t=0; t < combo_thread_data.thread_count; t++)
     {
         pthread_create(&thread[t], &attr, combo_threadfunc, &combo_thread_data);
@@ -876,9 +880,13 @@ int combo(int (*read_frame)(float *ref_data, float *main_data, float *temp_data,
     free_blur_buf(&combo_thread_data.ref_buf_array);
     free_blur_buf(&combo_thread_data.dis_buf_array);
     free_blur_buf(&combo_thread_data.blur_buf_array);
+
 #if BUF_OPT_ENABLE	
-	free_array(&motion_score_compute_flag_array);
+    free_array(&motion_score_compute_flag_array);
 #endif
+
+    free(thread);
+
     return 0;
 }